Rename Authenticator to PasswordAuthenticator. Other changes.
