+++ lib/pwauth/command/api.go
@@ -1,0 +1,30 @@
+package command
+
+import (
+	"github.com/Symantec/Dominator/lib/log"
+)
+
+type PasswordAuthenticator struct {
+	command string
+	args    []string
+	logger  log.Logger
+}
+
+// New creates a new PasswordAuthenticator. The command used to authenticate the
+// user is command and args may contain optional arguments to pass after the
+// username. Log messages are written to logger. A new *PasswordAuthenticator is
+// returned if the command exists, else an error is returned.
+func New(command string, args []string, logger log.Logger) (
+	*PasswordAuthenticator, error) {
+	return newAuthenticator(command, args, logger)
+}
+
+// PasswordAuthenticate will authenticate a user using the provided username and
+// password. The password is provided on the standard input of the
+// authentication command.
+// It returns true if the user is authenticated, else false (due to either
+// invalid username or incorrect password), and an error.
+func (pa *PasswordAuthenticator) PasswordAuthenticate(username string,
+	password []byte) (bool, error) {
+	return pa.passwordAuthenticate(username, password)
+}
